# Conflicts:
#	core/src/main/java/com/github/integritygame/objects/Bullet.java
#	core/src/main/java/com/github/integritygame/screens/MainGameScreen.java
#	core/src/main/java/com/github/integritygame/util/AssetManager.java
#	core/src/main/java/com/github/integritygame/util/GameManager.java

+++ core/src/main/java/com/github/integritygame/objects/PlayerHud.java
@@ -9,11 +9,11 @@ public class PlayerHud {
 
     private int x;
     private int y;
-    private int damage = 100;
+    private float damage = 100;
     private String name;
     private BitmapFont font;
     private int money = 100;
-    private int fuel = 100;
+    private float fuel = 100;
 
     private Tank tank;
